Reconcile new tests in api_tests into tests/api_tests
