+++ tests/api_tests/nodes/views/test_node_contributors_list.py
@@ -450,4 +450,3 @@ class TestNodeContributorAdd(NodeCRUDTestCase):
 
         self.private_project.reload()
         assert_not_in(self.user_two, self.private_project.contributors)
-
