+++ tests/api_tests/nodes/views/test_node_files_list.py
@@ -297,4 +297,3 @@ class TestNodeFilesListFiltering(ApiTestCase):
 
 
 
-
