+++ tests/api_tests/nodes/views/test_node_links_detail.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from nose.tools import *  # flake8: noqa
-
+from urlparse import urlparse
 from framework.auth.core import Auth
 
 from website.models import NodeLog
@@ -14,6 +14,8 @@ from tests.factories import (
     AuthUserFactory
 )
 from tests.utils import assert_logs
+
+node_url_for = lambda n_id: '/{}nodes/{}/'.format(API_BASE, n_id)
 
 
 class TestNodeLinkDetail(ApiTestCase):
@@ -40,14 +42,18 @@ class TestNodeLinkDetail(ApiTestCase):
         assert_equal(res.status_code, 200)
         assert_equal(res.content_type, 'application/vnd.api+json')
         res_json = res.json['data']
-        assert_equal(res_json['attributes']['target_node_id'], self.public_pointer_project._id)
+        expected_path = node_url_for(self.public_pointer_project._id)
+        actual_path = urlparse(res_json['relationships']['target_node']['links']['related']['href']).path
+        assert_equal(expected_path, actual_path)
 
     def test_returns_public_node_pointer_detail_logged_in(self):
         res = self.app.get(self.public_url, auth=self.user.auth)
         res_json = res.json['data']
         assert_equal(res.status_code, 200)
         assert_equal(res.content_type, 'application/vnd.api+json')
-        assert_equal(res_json['attributes']['target_node_id'], self.public_pointer_project._id)
+        expected_path = node_url_for(self.public_pointer_project._id)
+        actual_path = urlparse(res_json['relationships']['target_node']['links']['related']['href']).path
+        assert_equal(expected_path, actual_path)
 
     def test_returns_private_node_pointer_detail_logged_out(self):
         res = self.app.get(self.private_url, expect_errors=True)
@@ -59,12 +65,20 @@ class TestNodeLinkDetail(ApiTestCase):
         res_json = res.json['data']
         assert_equal(res.status_code, 200)
         assert_equal(res.content_type, 'application/vnd.api+json')
-        assert_equal(res_json['attributes']['target_node_id'], self.pointer_project._id)
+        expected_path = node_url_for(self.pointer_project._id)
+        actual_path = urlparse(res_json['relationships']['target_node']['links']['related']['href']).path
+        assert_equal(expected_path, actual_path)
 
     def test_returns_private_node_pointer_detail_logged_in_non_contributor(self):
         res = self.app.get(self.private_url, auth=self.user_two.auth, expect_errors=True)
         assert_equal(res.status_code, 403)
         assert_in('detail', res.json['errors'][0])
+
+    def test_self_link_points_to_node_link_detail_url(self):
+        res = self.app.get(self.public_url, auth=self.user.auth)
+        assert_equal(res.status_code, 200)
+        url = res.json['data']['links']['self']
+        assert_in(self.public_url, url)
 
 
 class TestDeleteNodeLink(ApiTestCase):
@@ -75,7 +89,7 @@ class TestDeleteNodeLink(ApiTestCase):
         self.project = ProjectFactory(creator=self.user, is_public=False)
         self.pointer_project = ProjectFactory(creator=self.user, is_public=True)
         self.pointer = self.project.add_pointer(self.pointer_project, auth=Auth(self.user), save=True)
-        self.private_url = '/{}nodes/{}/node_links/{}'.format(API_BASE, self.project._id, self.pointer._id)
+        self.private_url = '/{}nodes/{}/node_links/{}/'.format(API_BASE, self.project._id, self.pointer._id)
 
         self.user_two = AuthUserFactory()
 
@@ -84,7 +98,18 @@ class TestDeleteNodeLink(ApiTestCase):
         self.public_pointer = self.public_project.add_pointer(self.public_pointer_project,
                                                               auth=Auth(self.user),
                                                               save=True)
-        self.public_url = '/{}nodes/{}/node_links/{}'.format(API_BASE, self.public_project._id, self.public_pointer._id)
+        self.public_url = '/{}nodes/{}/node_links/{}/'.format(API_BASE, self.public_project._id, self.public_pointer._id)
+
+    def test_delete_node_link_no_permissions_for_target_node(self):
+        pointer_project = ProjectFactory(creator=self.user_two, is_public=False)
+        pointer = self.public_project.add_pointer(pointer_project, auth=Auth(self.user), save=True)
+        assert_in(pointer, self.public_project.nodes)
+        url = '/{}nodes/{}/node_links/{}/'.format(API_BASE, self.public_project._id, pointer._id)
+        res = self.app.delete_json_api(url, auth=self.user.auth, expect_errors=True)
+        assert_equal(res.status_code, 204)
+
+        self.public_project.reload()
+        assert_not_in(pointer, self.public_project.nodes)
 
     def test_cannot_delete_if_registration(self):
         registration = RegistrationFactory(project=self.public_project)
@@ -97,7 +122,7 @@ class TestDeleteNodeLink(ApiTestCase):
         assert_equal(res.status_code, 200)
         pointer_id = res.json['data'][0]['id']
 
-        url = '/{}nodes/{}/node_links/{}'.format(
+        url = '/{}nodes/{}/node_links/{}/'.format(
             API_BASE,
             registration._id,
             pointer_id,
@@ -169,7 +194,7 @@ class TestDeleteNodeLink(ApiTestCase):
         project = ProjectFactory(creator=self.user)
         # The node link belongs to a different project
         res = self.app.delete(
-            '/{}nodes/{}/node_links/{}'.format(API_BASE, project._id, self.public_pointer._id),
+            '/{}nodes/{}/node_links/{}/'.format(API_BASE, project._id, self.public_pointer._id),
             auth=self.user.auth,
             expect_errors=True
         )
